+++ example/storybook/src/styled/api/MultipleTheme/MultipleTheme.stories.tsx
@@ -1,0 +1,9 @@
+import type { ComponentMeta } from '@storybook/react-native';
+import { MultipleTheme } from './MultipleTheme';
+const MyDescendantsMeta: ComponentMeta<typeof MultipleTheme> = {
+  title: 'styled/api/stories/MultipleTheme',
+  component: MultipleTheme,
+};
+
+export { MultipleTheme } from './MultipleTheme';
+export default MyDescendantsMeta;